@@ -36,15 +36,15 @@ namespace hihapi.Models
         public String CreatorDisplayAs { get; set; }
 
         public ICollection<HomeMember> HomeMembers { get; set; }
-        public ICollection<FinanceAccountCategory> FinanceAccountCategories { get; set; }
-        public ICollection<FinanceAssetCategory> FinanceAssetCategories { get; set; }
-        public ICollection<FinanceDocumentType> FinanceDocumentTypes { get; set; }
-        public ICollection<FinanceTransactionType> FinanceTransactionTypes { get; set; }
-        public ICollection<FinanceAccount> FinanceAccounts { get; set; }
-        public ICollection<FinanceAccountExtraDP> FinanceAccountExtraDPs { get; set; }
-        public ICollection<FinanceAccountExtraAS> FinanceAccountExtraASs { get; set; }
-        public ICollection<FinanceControlCenter> FinanceControlCenters { get; set; }        
-        public ICollection<FinanceOrder> FinanceOrders { get; set; }
-        public ICollection<FinanceDocument> FinanceDocuments { get; set; }
+        //public ICollection<FinanceAccountCategory> FinanceAccountCategories { get; set; }
+        //public ICollection<FinanceAssetCategory> FinanceAssetCategories { get; set; }
+        //public ICollection<FinanceDocumentType> FinanceDocumentTypes { get; set; }
+        //public ICollection<FinanceTransactionType> FinanceTransactionTypes { get; set; }
+        //public ICollection<FinanceAccount> FinanceAccounts { get; set; }
+        //public ICollection<FinanceAccountExtraDP> FinanceAccountExtraDPs { get; set; }
+        //public ICollection<FinanceAccountExtraAS> FinanceAccountExtraASs { get; set; }
+        //public ICollection<FinanceControlCenter> FinanceControlCenters { get; set; }        
+        //public ICollection<FinanceOrder> FinanceOrders { get; set; }
+        //public ICollection<FinanceDocument> FinanceDocuments { get; set; }
     }
 }
