@@ -35,7 +35,7 @@ namespace hihapi.Models
         public const Int32 TranType_InsurancePaymentOut = 34;
 
         [Key]
-        [Column("ID", TypeName="int")]
+        [Column("ID", TypeName="INT")]
         public Int32 ID { get; set; }
 
         [Column("HID", TypeName="INT")]
